requirements.txt
fastapi[standard]
sqlalchemy
textblob
python-jose[cryptography] 
passlib[bcrypt==4.0.1] 
python-multipart